removing unecessary precision change

+++ app/src/main/java/com/dessalines/thumbkey/ui/components/keyboard/KeyboardKey.kt
@@ -78,7 +78,7 @@ fun KeyboardKey(
     capsLock: Boolean,
     legendSize: Int,
     keyPadding: Int,
-    keyBorderWidth: Float,
+    keyBorderWidth: Int,
     keyRadius: Int,
     minSwipeLength: Int,
     slideSensitivity: Int,
